+++ covasim/_v2_legacy/analysis.py
