+++ covasim/_v2_legacy/immunity.py
