+++ tests/_legacy/test_analysis.py
